backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.4.0
web3>=6.20.0
qrcode[pil]>=7.4.2
Pillow>=10.4.0
cryptography>=44.0.0
torch>=2.3.0
torch-geometric>=2.5.3
python-multipart>=0.0.9
